+++ app/javascript/packs/attendanceManagementPage/AttendanceManagementPage.jsx
@@ -1,8 +1,6 @@
 import React, { useState, useEffect } from 'react'
 import axios from 'axios'
-import useCookies from '../general/useCookies'
 import { NavigationBar } from '../general/NavigationBar'
-import { handleServerError } from '../general/handleServerError'
 import { HandleDownloadWithExcelJS } from './AnnualAttendanceExcel'
 import { ParadeList } from './ParadeList'
 import { NewParadeForm } from './NewParadeForm'
@@ -10,7 +8,6 @@ import { ParadeInformation } from './ParadeInformation'
 
 // To access attendance records and take new attendance
 const AttendanceManagementPage = () => {
-  const cookies = useCookies()
   const [pageState, setPageState] = useState('list')
   const [renderPage, setRenderPage] = useState(false)
   const [reload, setReload] = useState(false)
@@ -41,9 +38,8 @@ const AttendanceManagementPage = () => {
         <div className='annual-attendance-list'>
           <h1>Yearly Attendance File</h1>
           {years.map((year) => {
-            return(<HandleDownloadWithExcelJS year={year}/>)
+            return(<HandleDownloadWithExcelJS key={year} year={year}/>)
           })}
-          {pageState != 'list' && pageState != 'form' && pageState.includes('Y') && <AnnualAttendanceExcel year={pageState.split('Y')[1]}/>}
         </div>
         <br/>
         <ParadeList reload={reload} setPageState={setPageState}/>